Actual introduction date of gununu was on 2011-07-26.
As far as I can tell, there is nothing significant about 2012-12-16.
The original release of commentfaces (using the `/` syntax) came on
2011-07-26:
https://old.reddit.com/r/anime/comments/izxos/comment_faces_for_ranime/

The commentface refresh to use the new syntax and introduction to
several more faces occurred on 2015-02-13, but gununu and 17 others
predate that and consequently the introduction date should reflect that
accordingly.