somad
The actual date for the early 2016 comment face removal is uncertain, so
I decided to move forward with 2011-03-21 for the following reasons:

1) The comment faces page
(https://old.reddit.com/r/anime/wiki/commentfaces) says early 2016.
2) The retired comment faces page which contains the data for the
screencap for the comment faces page was last updated on 2016-03-14.
3) Archives from the wayback machine on and immediately after that date
still contain the old comment faces.  The earliest archive I can get to
resolve from there with those faces removed is 2016-03-29.  The last
archive I can get to resolve from there still containing those faces is
2016-03-18.
4) The removed date is after the date that the new faces were announced
and presumably provided (2016-03-21 -
https://old.reddit.com/r/anime/comments/4babw1/new_comment_faces/).
5) I'm presuming that the announcement date for the new faces aligns
with the removal date for the old ones in order to have space for the
new ones in the first place and also to provide the smoothest transition
between them.